laragig completed

+++ app/Http/Controllers/ListingController.php
@@ -5,24 +5,32 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Validation\Rule;
 
 class ListingController extends Controller
 {
     // @desc show all listing
-    public function index()
+    public function index(Listing $listing)
     {
-        echo Auth::id();
+
+        $userId = Auth::id();
+
         return view("listings.index", [
-            "listings" => Listing::latest()->filter(request(["tag", "search"]))->paginate(2)
+            "listings" => Listing::latest()->filter(request(["tag", "search"]))->paginate(2),
+
+
         ]);
     }
 
     // @desc show single listing
     public function show(Listing $listing)
     {
+        $userId = Auth::id();
+
         return view("listings.show", [
-            "listing" => $listing
+            "listing" => $listing,
+            "isOwner" => $listing->user_id === $userId
         ]);
     }
 
@@ -31,8 +39,23 @@ class ListingController extends Controller
         return view("listings.create");
     }
 
+    public function manage()
+    {
+        $userId = Auth::id();
+        $userListings = DB::select('SELECT * FROM listings WHERE user_id = ?', [$userId]);
+
+        return view("listings.manage", [
+            "listings" => $userListings
+        ]);
+    }
+
     public function edit(Listing $listing, Request $request)
     {
+        if ($listing->user_id !== Auth::id()) {
+            abort(403, "Unauthorized");
+            return;
+        }
+
         return view("listings.edit", [
             "listing" => $listing
         ]);
@@ -55,6 +78,8 @@ class ListingController extends Controller
             $formFileds["logo"] = $request->file("logo")->store("logos", "public");
         }
 
+        $formFileds["user_id"] = Auth::id();
+
         Listing::create($formFileds);
 
         return redirect("/", 201)->with("success", "Successfully created");
@@ -62,6 +87,12 @@ class ListingController extends Controller
 
     public function update(Listing $listing, Request $request)
     {
+
+        if ($listing->user_id !== Auth::id()) {
+            abort(403, "Unauthorized");
+            return;
+        }
+
         $formFileds = $request->validate([
             "title" => "required",
             "tags" => "required",
@@ -83,6 +114,11 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id !== Auth::id()) {
+            abort(403, "Unauthorized");
+            return;
+        }
+
         $listing->delete();
 
         return redirect("/")->with("message", "Listing Delete Successfully");
